REFACTOR(componetes/mentor-register.. & forms..): refatorando componente de Loading
