+++ src/pages/user/components/mentor-register/styled.js
@@ -59,9 +59,7 @@ background-color: rgb(250, 128, 114);
 text-align: center;
 color: rgb(139,0,0)
 `
-export const Loading = styled.div`
-  text-align: center;
-`
+
 export const MessageErrorPassword = styled.div`
 font-size: 16px;
 color: #ff3d00;
